refactor(front): StrategyCard에서 StrategySections 분리
